pyproject.toml
[project]
name = "auto-bing-rewards"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles==23.2.1",
    "apscheduler>=3.11.2",
    "fastapi==0.115.0",
    "loguru>=0.7.3",
    "playwright==1.48.0",
    "playwright-stealth==2.0.1",
    "pydantic==2.9.2",
    "pydantic-settings>=2.12.0",
    "python-multipart==0.0.12",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "sqlalchemy==2.0.35",
    "uvicorn==0.32.0",
]
